Step 2: the design triggers warnings (#5649)
* changing design to remove console warnings

* update changelog

* Changes in the display of pop-up windows in the agents log section

* semicolon is added
