+++ plugins/main/public/controllers/register-agent/components/group-input/group-input.scss
@@ -1,7 +1,4 @@
-.groupTitle {
-  margin-top: '32px';
-  flex-direction: 'row';
-  font-style: normal;
+.registerAgentLabels {
   font-weight: 700;
   font-size: 12px;
   line-height: 20px;
